moving artifact mocks to shared location
