+++ Tasks/Common/packaging-common/Tests/ArtifactToolMockHelper.ts
@@ -1,0 +1,33 @@
+import tmrm = require('vsts-task-lib/mock-run');
+
+export function registerArtifactToolUtilitiesMock(tmr: tmrm.TaskMockRunner, toolPath: string) {
+    tmr.registerMock('packaging-common/universal/ArtifactToolUtilities', {
+        getArtifactToolFromService: function(serviceUri, accessToken, toolName) {
+            return toolPath;
+        },
+        getPackageNameFromId: function(serviceUri: string, accessToken: string, feedId: string, packageId: string) {
+            return packageId;
+        }
+    });
+}
+
+export function registerArtifactToolRunnerMock(tmr: tmrm.TaskMockRunner) {
+    var mtt = require('vsts-task-lib/mock-toolrunner');
+    tmr.registerMock('packaging-common/universal/ArtifactToolRunner', {
+        getOptions: function() {
+            return {
+                cwd: process.cwd(),
+                env: Object.assign({}, process.env),
+                silent: false,
+                failOnStdErr: false,
+                ignoreReturnCode: false,
+                windowsVerbatimArguments: false
+            }
+        },
+        runArtifactTool: function(artifactToolPath: string, command: string[], execOptions) {
+            var tr = new mtt.ToolRunner(artifactToolPath)
+            tr.arg(command);
+            return tr.execSync(execOptions);
+        }
+    });
+}
